+++ backend/ormconfig.json
@@ -7,12 +7,12 @@
   "database": "faq_support",
   "logging": true,
   "entities": [
-    "./src/models/*.ts"
+    "./src/modules/**/infra/typeorm/entities/*.ts"
   ],
   "migrations": [
-    "./src/database/migrations/*.ts"
+    "./src/shared/infra/database/migrations/*.ts"
   ],
   "cli": {
-    "migrationsDir": "/src/database/migrations/"
+    "migrationsDir": "./src/shared/infra/database/migrations/"
   }
 }
