--- backend/src/models/Category.ts
@@ -1,23 +1,0 @@
-import { Entity, PrimaryGeneratedColumn, Column, CreateDateColumn, OneToMany, UpdateDateColumn } from 'typeorm';
-
-import Answer from './Answer';
-
-@Entity('categories')
-class Category {
-	@PrimaryGeneratedColumn('uuid')
-	id: string;
-
-	@Column()
-	name: string;
-
-	@OneToMany(() => Answer, answer => answer.category)
-	answers: Answer;
-
-	@CreateDateColumn()
-	created_at: Date;
-
-	@UpdateDateColumn()
-	updated_at: Date;
-}
-
-export default Category;
