+++ backend/src/modules/answers/infra/typeorm/entities/Category.ts
@@ -1,0 +1,30 @@
+import {
+	Entity,
+	PrimaryGeneratedColumn,
+	Column,
+	CreateDateColumn,
+	OneToMany,
+	UpdateDateColumn,
+} from 'typeorm';
+
+import Answer from './Answer';
+
+@Entity('categories')
+class Category {
+	@PrimaryGeneratedColumn('uuid')
+	id: string;
+
+	@Column()
+	name: string;
+
+	@OneToMany(() => Answer, answer => answer.category)
+	answers: Answer;
+
+	@CreateDateColumn()
+	created_at: Date;
+
+	@UpdateDateColumn()
+	updated_at: Date;
+}
+
+export default Category;
