+++ backend/src/modules/answers/services/CreateAnswerService.ts
@@ -1,0 +1,47 @@
+import { getRepository } from 'typeorm';
+
+import Answer from '@modules/answers/infra/typeorm/entities/Answer';
+import Category from '@modules/answers/infra/typeorm/entities/Category';
+
+import AppError from '@shared/errors/AppError';
+
+interface Request {
+	title: string;
+	reply: string;
+	category?: string;
+}
+
+class CreateUser {
+	public async execute({ reply, category, title }: Request): Promise<Answer> {
+		const answerRepository = getRepository(Answer);
+		const categoryRepository = getRepository(Category);
+
+		const findCategoryId = await categoryRepository.findOne({
+			where: { name: category },
+		});
+
+		const findExactAnswer = await answerRepository.findOne({
+			where: { title, reply },
+		});
+
+		if (!findCategoryId) {
+			throw new AppError('Category not registered');
+		}
+
+		if (findExactAnswer) {
+			throw new AppError('This answer is already exist');
+		}
+
+		const newAnswer = answerRepository.create({
+			title,
+			reply,
+			category_id: findCategoryId.id,
+		});
+
+		await answerRepository.save(newAnswer);
+
+		return newAnswer;
+	}
+}
+
+export default CreateUser;
