+++ backend/src/modules/answers/services/DeleteAnswerService.ts
@@ -1,0 +1,11 @@
+// import { getRepository } from 'typeorm';
+
+// import Issue from '../models/Issue';
+
+class DeleteIssues {
+	public async execute(): Promise<void> {
+		//
+	}
+}
+
+export default DeleteIssues;
