+++ backend/src/modules/issues/services/DeleteIssueService.ts
@@ -1,0 +1,11 @@
+// import { getRepository } from 'typeorm';
+
+// import Issue from '@modules/issues/infra/typeorm/entities/Issue';
+
+class DeleteIssues {
+	public async execute(): Promise<void> {
+		//
+	}
+}
+
+export default DeleteIssues;
