+++ backend/src/modules/users/infra/http/routes/sessions.routes.ts
@@ -1,0 +1,11 @@
+import { Router } from 'express';
+
+import SessionsController from '@modules/users/infra/http/controllers/SessionsController';
+
+const sessionRouter = Router();
+
+const sessionsController = new SessionsController();
+
+sessionRouter.post('/', sessionsController.create);
+
+export default sessionRouter;
