+++ backend/src/modules/users/infra/typeorm/entities/User.ts
@@ -1,0 +1,30 @@
+import {
+	Entity,
+	Column,
+	PrimaryGeneratedColumn,
+	CreateDateColumn,
+	UpdateDateColumn,
+} from 'typeorm';
+
+@Entity('users')
+class Users {
+	@PrimaryGeneratedColumn('uuid')
+	id: string;
+
+	@Column()
+	name: string;
+
+	@Column()
+	email: string;
+
+	@Column()
+	password: string;
+
+	@CreateDateColumn()
+	created_at: Date;
+
+	@UpdateDateColumn()
+	updated_at: Date;
+}
+
+export default Users;
