--- backend/src/routes/sessions.routes.ts
@@ -1,21 +1,0 @@
-import { Router } from 'express';
-import AuthenticateUserService from '../service/AuthenticateUserService';
-
-const sessionRouter = Router();
-
-sessionRouter.post('/', async (request, response) => {
-	const { email, password } = request.body;
-
-	const authenticateUser = new AuthenticateUserService();
-
-	const { user, token } = await authenticateUser.execute({
-		email, password
-	});
-
-	delete user.password;
-
-	return response.json({user, token});
-});
-
-
-export default sessionRouter;
