--- backend/src/server.ts
@@ -1,37 +1,0 @@
-import 'reflect-metadata';
-
-import express, {Request, Response, NextFunction} from 'express';
-import cors from 'cors';
-
-import 'express-async-errors';
-
-import routes from './routes';
-import AppError from './errors/AppError';
-
-import './database';
-
-const app = express();
-
-app.use(express.json());
-app.use(cors());
-app.use(routes);
-
-app.use((err: Error, request: Request, response: Response, _: NextFunction) => {
-  if(err instanceof AppError) {
-    return response.status(err.statusCode).json({
-      status: 'error',
-      message: err.message,
-    });
-  }
-
-  console.error(err);
-
-  return response.status(500).json({
-    status: 'error',
-    message: 'Internal server error',
-  });
-});
-
-app.listen(3333, () => {
-  console.log('🚀 Server is running on port 3333 🚀');
-});
