--- backend/src/service/AuthenticateUserService.ts
@@ -1,52 +1,0 @@
-import { getRepository } from 'typeorm';
-import { compare } from 'bcryptjs';
-import { sign } from 'jsonwebtoken';
-import authConfig from '../config/auth';
-
-import AppError from '../errors/AppError';
-
-import User from '../models/User';
-
-interface Request{
-	email: string;
-	password: string;
-}
-
-interface Response{
-	user: User;
-	token: string;
-}
-
-class AuthenticateUserService {
-	public async execute({email, password}: Request): Promise<Response> {
-		const userRepository = getRepository(User);
-
-		 const user = await userRepository.findOne({
-			where: {email}
-		})
-
-		if(! user){
-			throw new AppError('invalid login credentials', 401)
-		}
-
-		const passwordCompare = await compare(password, user.password);
-
-		if (! passwordCompare){
-			throw new AppError('invalid login credentials', 401);
-		}
-
-		const {secret, expiresIn} = authConfig.jwt;
-
-		const token = sign({}, secret, {
-			subject: user.id,
-			expiresIn,
-			});
-
-		return {
-			user,
-			token,
-		}
-	}
-}
-
-export default AuthenticateUserService;
