+++ backend/src/shared/errors/AppError.ts
@@ -1,0 +1,12 @@
+class AppError {
+	public readonly message: string;
+
+	public readonly statusCode: number;
+
+	constructor(message: string, statusCode = 400) {
+		this.message = message;
+		this.statusCode = statusCode;
+	}
+}
+
+export default AppError;
